requirements.txt
streamlit
gspread
google-auth
pandas
plotly
pillow
requests
python-dotenv
ultralytics
statsmodels
